requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.9
django-tastypie==0.14.0
djangorestframework==3.7.7
ez-setup==0.9
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
psycopg2==2.7.3.2
Pygments==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
